+++ aupay-user/src/main/java/com/aucloud/aupay/user/orm/po/AccountAssetsRecord.java
@@ -32,12 +32,12 @@ public class AccountAssetsRecord implements Serializable {
      * 主键id
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 资产表id
      */
-    private Integer assetsId;
+    private Long assetsId;
 
     /**
      * 交易类型
